+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Development Version:
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.
 - Create/Attach dimension scales on append (``mode="r+"``)
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.
+- Create/Attach/Detach dimension scales only if necessary.
+  By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.
 
 Version 0.11.0 (April 20, 2021):